+++ changelog.md
@@ -1,5 +1,29 @@
 # The Modding Tree changelog:
 
+### v2.1.1 - 10/17/20
+- Added resource-display component, which displays the base currency for the prestige layer, as well as the best
+    and/or total of this layer's prestige currency.
+- Fixed the value for the base currency not updating in resource-display.
+
+## v2.1: We should have thought of this sooner! - 10/17/20
+- Moved most of the code users will want to edit to mod.js, added documentation for it.
+    - Specifically, modInfo, VERSION, canGenPoints, getPointGen, and maxTickLength
+- Added getStartPoints()
+- Added the ability to store non-layer-related data
+- Added the ability to display more things at the top of the tree tab below points.
+- Made the endgame condition customizable
+- Added "sell one" and "sell all" buttons for buyables.
+- Moved the old "game" to demo.js, and replaced it with a minimal game that won't cause issues when edited.
+- Fixed issues with version number
+- Fixed number formatting issue making things like "10e9" appear.
+
+
+### v2.0.5 - 10/16/20
+- Made more features (including prestige parameters) able to be dynamic.
+- Layer nodes can be hidden but still take up space with "ghost" visibility
+- Added clickableEffect for real.
+- Fixed some visual issues with bars.
+- A few other minor tweaks and improvements.
 
 ### v2.0.4 - 10/16/20
 - Fixed HTML on buttons interfering with clicking on them.